data_ingestion/language_agent/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
spacy>=3.0.0
textblob>=0.15.3
python-dotenv>=0.19.0
langdetect>=1.0.9
pydantic>=1.8.0
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
accelerate>=0.20.0 